feat(#34):editar perfil y eventos de usuario solo confirmados

+++ database/migrations/2020_01_22_120004_create_carritos_table.php
@@ -19,7 +19,7 @@ class CreateCarritosTable extends Migration
             $table->integer('usuario_id')->unsigned();
             $table->foreign('usuario_id')->references('id')->on('users');
             $table->boolean('confirmado');
-            $table->integer('precio');
+            $table->decimal('precio', 6, 2);
         });
     }
